Day-12-16/E-Commerce-FASTApi/requirements.txt
apscheduler==3.11.0
Faker==25.2.0
fastapi==0.115.13
pydantic==2.11.7
python-dotenv==1.1.0
SQLAlchemy==2.0.41
uvicorn==0.30.1
pymysql==1.1.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
requests==2.31.0
tqdm
httx
tqdm